Update to reflect removal of SincFlux in favor of CircularApertureFlux
The changes in the 'expected results' files are due to a change from a default
radius of 7 to 6 (pixels) when we switched algorithms.  Before updating the
files, I also verified that if we run with a 7 pixel aperture with the new code,
the results match the old algorithm (aside from round-off error difference due
to a slightly different implementation).