Update Linux64 expected files as of DM-4836
The update in DM-4836 fixes a bug in the flow of aperture corrections
which had the correction being applied after all plugins had run,
independent of their execution order.  This resulted in plugins whose
measurements rely on aperture corrected fluxes (i.e. with execution
order > APCORR_ORDER) being applied prior to the aperture correction,
leading to erroneous results. The only plugin currently affected by
this is base_ClassificationExtendedness.

This updates the expected values in the demo to the correct ones
after applying the DM-4836 fix.