Send log to stderr
The default configuration of lsst.log is at DEBUG level sent to
stdout; the default of lsst.pex.logging is at INFO level sent
to stderr.  As the stack is converted to use lsst.log instead of
lsst.pex.logging, it would add extra lines to the outputted
detected-sources file, and fail the comparison.  This commit
configures the log to be similar to the old pex_logging.