Update Linux64 expected files as of DM-6815
In https://jira.lsstcorp.org/browse/DM-6815 we update various
config settings according to lessons learned in HSC production
runs.  This results in small changes to the output values for
a given source, but there is a significant change to the number
of output sources.  The latter is due to the activation of junk
suppression (see https://jira.lsstcorp.org/browse/RFC-212) which
greatly reduces the number of "junk" detections around bright
objects (see Jira ticket for demonstrative figures).

This updates the expected files for Linux64 after applying the
default settings of DM-6815.