Update expected output as of DM-8044
In DM-8044 we turn off the junk suppression in coadd assembly and single frame
processing in order to obtain the desired behaviour for safe clipping coadd
assembly.  This results in small changes in the behaviour of image
characterization, detection, and measurement.

This updates the expected files for both DarwinX86 and Linux64 after applying
the default settings of DM-8044.